+++ Ads.h
@@ -28,9 +28,9 @@ class Ads {
 	~Ads();		// destructor
 
         bool getAds ( char           *q         , 
-                      long            qlen      ,
-                      long            pageNum   ,
-                      long            queryIP   ,
+                      int32_t            qlen      ,
+                      int32_t            pageNum   ,
+                      int32_t            queryIP   ,
                       char           *coll      ,
                       void           *state     ,
                       void          (*callback)(void *state) );
@@ -39,8 +39,8 @@ class Ads {
 	// . "q" is the NULL terminated query
 	// . "ip" is the ip of the user performing the query
 		
-	long  getNumPaidInclusionAds ( ) { return m_numAds[m_indexPIAds]; };
-	long  getNumSkyscraperAds    ( ) { return m_numAds[m_indexSSAds]; };
+	int32_t  getNumPaidInclusionAds ( ) { return m_numAds[m_indexPIAds]; };
+	int32_t  getNumSkyscraperAds    ( ) { return m_numAds[m_indexSSAds]; };
         bool  hasAds                 ( ) { 
                 return (bool)(m_numAds[m_indexPIAds] || m_numAds[m_indexSSAds]);
         };
@@ -48,64 +48,64 @@ class Ads {
                 return (bool)(m_numGotAds >= MAX_FEEDS);
         };
 
-        void  printPaidInclusionAds(SafeBuf *sb, long numCharPerLine );
-        void  printSkyscraperAds   (SafeBuf *sb, long numCharPerLine );
+        void  printPaidInclusionAds(SafeBuf *sb, int32_t numCharPerLine );
+        void  printSkyscraperAds   (SafeBuf *sb, int32_t numCharPerLine );
 
 	// need to keep public so wrapper can call
-        void  gotDoc( class TcpSocket *ts, long index );
+        void  gotDoc( class TcpSocket *ts, int32_t index );
         void  selectDisplayAds( );
 	void   *m_state;
 	void  (*m_callback)(void *state);
 
         //static void initCollAvailAds(              );
         //static void setAvailableAds ( char *coll   );
-        //static long getAdFeedIndex  ( collnum_t cn );
+        //static int32_t getAdFeedIndex  ( collnum_t cn );
 
  private:
-        bool getAd ( long            index     ,
+        bool getAd ( int32_t            index     ,
                      char           *cgi       ,
-                     long            numAds    );
+                     int32_t            numAds    );
         void printAd( SafeBuf *sb    , 
-                      char *url      , long urlLen,
-                      char *title    , long titleLen,
-                      char *desc     , long descLen,
-                      char *site     , long siteLen,
-                      long  numCharPerLine );
+                      char *url      , int32_t urlLen,
+                      char *title    , int32_t titleLen,
+                      char *desc     , int32_t descLen,
+                      char *site     , int32_t siteLen,
+                      int32_t  numCharPerLine );
 
         char *m_coll;
-        long  m_queryIP;      
-        long  m_pageAds;
-        long  m_feedIndex;
-        long  m_indexPIAds;
-        long  m_indexSSAds;
-        long  m_numGotAds;
+        int32_t  m_queryIP;      
+        int32_t  m_pageAds;
+        int32_t  m_feedIndex;
+        int32_t  m_indexPIAds;
+        int32_t  m_indexSSAds;
+        int32_t  m_numGotAds;
         bool  m_adSSSameasPI;
         bool  m_adBSSSameasBPI;
         CollectionRec *m_cr;
 
         char  m_q        [ MAX_AD_QUERY_LEN ];
-        long  m_qlen;
+        int32_t  m_qlen;
 	char *m_titles   [MAX_FEEDS][ MAX_ADS ];
-	long  m_titlesLen[MAX_FEEDS][ MAX_ADS ];
+	int32_t  m_titlesLen[MAX_FEEDS][ MAX_ADS ];
 	char *m_desc     [MAX_FEEDS][ MAX_ADS ];
-	long  m_descLen  [MAX_FEEDS][ MAX_ADS ];
+	int32_t  m_descLen  [MAX_FEEDS][ MAX_ADS ];
 	char *m_sites    [MAX_FEEDS][ MAX_ADS ];
-	long  m_sitesLen [MAX_FEEDS][ MAX_ADS ];
+	int32_t  m_sitesLen [MAX_FEEDS][ MAX_ADS ];
 	char *m_urls     [MAX_FEEDS][ MAX_ADS ];
-	long  m_urlsLen  [MAX_FEEDS][ MAX_ADS ];
-	long  m_numAds   [MAX_FEEDS];
+	int32_t  m_urlsLen  [MAX_FEEDS][ MAX_ADS ];
+	int32_t  m_numAds   [MAX_FEEDS];
 	char *m_buf      [MAX_FEEDS];
-	long  m_bufLen   [MAX_FEEDS];
+	int32_t  m_bufLen   [MAX_FEEDS];
 	//Url m_url      [MAX_FEEDS];
 	Xml   m_xml      [MAX_FEEDS];
 
-        //static long s_availableAds[16][MAX_AD_FEEDS];
-        //static long s_numAvailableAds[16];
+        //static int32_t s_availableAds[16][MAX_AD_FEEDS];
+        //static int32_t s_numAvailableAds[16];
 };
 
 struct AdFeed {
         Ads  *m_ads;
-        long  m_index;
+        int32_t  m_index;
 };
 
 #endif
