+++ Blaster.h
@@ -14,16 +14,16 @@ struct StateBD {
 	char *m_u1,*m_u2;
 	char *m_u;
 	char *m_buf1;
-	long m_buf1Len;
-	long m_buf1MaxLen;
-	long m_numUrlDocsSent;
-	long m_numUrlDocsReceived;
+	int32_t m_buf1Len;
+	int32_t m_buf1MaxLen;
+	int32_t m_numUrlDocsSent;
+	int32_t m_numUrlDocsReceived;
 	SafeBuf m_injectUrl;
 };
 
 struct StateBD2 {
 	StateBD *m_st;
-	long m_numRedirects;
+	int32_t m_numRedirects;
 	//Url m_url;
 	char *m_url;
 };
@@ -38,7 +38,7 @@ class Blaster {
 	bool init();
 
 	void runBlaster(char *file1,char *file2,
-			long maxNumThreads, long wait, 
+			int32_t maxNumThreads, int32_t wait, 
 			bool isLogFile,
 			bool verbose,bool justDisplay,
 			bool useProxy,
@@ -63,26 +63,26 @@ class Blaster {
 	bool m_doInjectionWithLinks;
 
  protected:
-	long  m_maxNumThreads;
-	long  m_launched;
+	int32_t  m_maxNumThreads;
+	int32_t  m_launched;
 	char *m_buf1;
-	long  m_bufSize1;
+	int32_t  m_bufSize1;
 	char *m_buf2;
-	long m_bufSize2;
+	int32_t m_bufSize2;
 	char *m_p1;
 	char *m_p1end;
 	char *m_p2;
 	char *m_p2end;
-	long  m_totalUrls;
-	long  m_totalDone;
+	int32_t  m_totalUrls;
+	int32_t  m_totalDone;
 	bool  m_portSwitch;
-	long  m_wait;
+	int32_t  m_wait;
 	bool  m_justDisplay;
 	bool  m_useProxy;
 	bool  m_verbose;
 	bool  m_isLogFile;
 	int64_t m_startTime;
-	long m_numUrls;
+	int32_t m_numUrls;
 	char **m_lineStart;
 	bool m_blasterDiff;
 	bool m_print;
